Move schema-command.js so it doesn't show up in help
