+++ src/commands/schema/ls.js
@@ -1,4 +1,4 @@
-const SchemaCommand = require("./schema-command.js");
+const SchemaCommand = require("../../lib/schema-command.js");
 const fetch = require("node-fetch");
 
 class ListSchemaCommand extends SchemaCommand {
